cleaned up game.ts
