+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@babylonjs/core": "^6.31.0",
     "@babylonjs/gui": "^6.35.0",
+    "@babylonjs/havok": "^1.3.1",
     "@babylonjs/inspector": "^6.31.0",
     "@babylonjs/loaders": "^6.31.0",
     "alea": "^1.0.1",
